added oily_png for faster processing, added recompose.rb for composing the final image

+++ generate_blender_sequence_images.rb
@@ -5,10 +5,12 @@ require 'csv'
 require './lib/libdejong'
 require 'fileutils'
 
-options = {}
+options = {
+  :keyframesonly => false
+}
 
 OptionParser.new do |opts|
-  opts.banner = "Usage: convert_wii.rb --data [directory] --out [directory] --tag [tag] --file [filename]"
+  opts.banner = "Usage: convert_wii.rb --data [directory] --out [directory] --tag [tag] --red [255] --green [255] --blue [255] --file [filename]"
 
   opts.on("--data [ARG]", "Input Data") do |v|
     options[:data] = v
@@ -22,12 +24,29 @@ OptionParser.new do |opts|
   opts.on("--tag [ARG]", "output file tag") do |v|
     options[:tag] = v
   end
+  opts.on("--red [ARG]", "red") do |v|
+    options[:red] = v.to_i
+  end
+  opts.on("--green [ARG]", "green") do |v|
+    options[:green] = v.to_i
+  end
+  opts.on("--blue [ARG]", "blue") do |v|
+    options[:blue] = v.to_i
+  end
+  opts.on("--keyframes-only", "only generate keyframes") do |v|
+    options[:keyframesonly] = true
+  end
+
 
 end.parse!
 
 
 file = options[:file] 
 seq = 0
+
+r = options[:red]
+g = options[:green]
+b = options[:blue]
 
 
 CSV.foreach(file, :col_sep => ",") do |row|
@@ -37,7 +56,7 @@ CSV.foreach(file, :col_sep => ",") do |row|
   p = LibDeJong::Point.new x, y
   fname = File.join( options[:data], "dejong-1024-#{p.to_filename_part}.json.gz" )
   background = ChunkyPNG::Color::TRANSPARENT
-  color = ChunkyPNG::Color.rgb(0,0,0)
+  color = ChunkyPNG::Color.rgb(r,g,b)
 
   if File.exists? fname 
     imagename = "dejong-1024-#{p.to_filename_part}-#{ChunkyPNG::Color.to_hex(color, false)}.png"
@@ -51,11 +70,11 @@ CSV.foreach(file, :col_sep => ",") do |row|
     else
       if !File.exists? imagenamedir
         frame = LibDeJong::FrameWriter.read_raw( fname )
-        LibDeJong::FrameWriter.write frame, options[:output], background, 255, imagename
-        puts "wrote #{imagename}"
+        LibDeJong::FrameWriter.write frame, options[:output], background, color, imagename
+        puts "wrote [#{seq}] #{options[:tag]} #{imagename}"
       end
 
-      if File.exists? imagenamedir
+      if File.exists?(imagenamedir) && !options[:keyframesonly]
         #copy imagename to seqname
         FileUtils.cp imagenamedir, seqnamedir
       else
